refactor: remove ToolUseCycle.ts wrapper file
Delete runToolUseCycle() and ToolUseCycle.ts entirely:
- Update imports to use ToolUseCycleOptions from CycleServices directly
- No more wrapper layers for tool-use cycle execution

Files updated:
- ToolUseServices.ts, ToolUseRunFlow.ts, ToolUseFlowContext.ts, BashTool.test.ts
