+++ src/agent/implementations/flows/tooluse/ToolUseServices.ts
@@ -10,7 +10,7 @@
  * dependencies (tool registry, session, cycle operations).
  */
 
-import type { ToolUseCycleOptions } from '@agent/core/ToolUseCycle';
+import type { ToolUseCycleOptions } from '@agent/core/flows/CycleServices';
 import type { AgentSharedStore } from '@agent/core/AgentSharedStore';
 import type { AgentRunState } from '@agent/core/AgentState';
 import type { AgentToolUseSetting } from '@agent/core/AgentDataclass';
